# lib/api-zod/src/generated/types/approvalStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Recipe Nutrition Manager API
 * OpenAPI spec version: 0.1.0
 */

export type ApprovalStatus = typeof ApprovalStatus[keyof typeof ApprovalStatus];


export const ApprovalStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
} as const;