# lib/api-zod/src/generated/types/cookingLog.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Recipe Nutrition Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface CookingLog {
  id: bigint;
  user: User;
  note: string;
  cookedAt: Date;
}