# lib/api-zod/src/generated/types/cookingLogInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Recipe Nutrition Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface CookingLogInput {
  /** @minLength 1 */
  note: string;
  /** @nullable */
  cookedAt?: Date | null;
}