# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Recipe Nutrition Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Recipe } from './recipe';

export interface DashboardSummary {
  recipeCount: number;
  ingredientCount: number;
  pendingApprovalCount: number;
  averageRating: number;
  recentRecipes: Recipe[];
}