# lib/api-zod/src/generated/types/ingredientInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Recipe Nutrition Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface IngredientInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 0 */
  caloriesPerUnit: number;
  /** @minLength 1 */
  unit: string;
  alternativeIds?: bigint[];
}