# lib/api-zod/src/generated/types/ingredientReference.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Recipe Nutrition Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface IngredientReference {
  id: bigint;
  name: string;
}