# lib/api-zod/src/generated/types/ratingInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Recipe Nutrition Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface RatingInput {
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
}