# lib/api-zod/src/generated/types/recipeIngredient.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Recipe Nutrition Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { RecipeIngredientInput } from './recipeIngredientInput';

export type RecipeIngredient = RecipeIngredientInput & {
  ingredientName: string;
  unit: string;
  calories: number;
};